+++ lib/feature/kamoku_detail/kamoku_detail_kakomon_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:dotto/components/s3.dart';
-import 'package:dotto/components/widgets/progress_indicator.dart';
+import 'package:dotto/widget/progress_indicator.dart';
 import 'package:dotto/feature/kamoku_detail/repository/kamoku_detail_repository.dart';
 import 'package:dotto/feature/kamoku_detail/widget/kamoku_detail_kakomon_list_objects.dart';
 
@@ -10,22 +10,18 @@ class KamokuDetailKakomonListScreen extends StatefulWidget {
   final int url;
 
   @override
-  State<KamokuDetailKakomonListScreen> createState() =>
-      _KamokuDetailKakomonListScreenState();
+  State<KamokuDetailKakomonListScreen> createState() => _KamokuDetailKakomonListScreenState();
 }
 
-class _KamokuDetailKakomonListScreenState
-    extends State<KamokuDetailKakomonListScreen> {
+class _KamokuDetailKakomonListScreenState extends State<KamokuDetailKakomonListScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Center(
         child: (KamokuDetailRepository().isLoggedinGoogle())
             ? FutureBuilder(
-                future:
-                    S3.instance.getListObjectsKey(url: widget.url.toString()),
-                builder: (BuildContext context,
-                    AsyncSnapshot<List<String>> snapshot) {
+                future: S3.instance.getListObjectsKey(url: widget.url.toString()),
+                builder: (BuildContext context, AsyncSnapshot<List<String>> snapshot) {
                   if (snapshot.hasData) {
                     if (snapshot.data!.isEmpty) {
                       return const Center(
