+++ lib/feature/kamoku_search/widget/kamoku_search_result.dart
@@ -1,6 +1,6 @@
 import 'package:dotto/importer.dart';
 import 'package:dotto/components/animation.dart';
-import 'package:dotto/components/widgets/progress_indicator.dart';
+import 'package:dotto/widget/progress_indicator.dart';
 import 'package:dotto/feature/kamoku_detail/kamoku_detail_page_view.dart';
 import 'package:dotto/feature/kamoku_search/controller/kamoku_search_controller.dart';
 import 'package:dotto/feature/kamoku_search/repository/kamoku_search_repository.dart';
