+++ lib/feature/my_page/feature/timetable/course_cancellation.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:dotto/controller/user_controller.dart';
 import 'package:dotto/feature/my_page/feature/timetable/controller/timetable_controller.dart';
 import 'package:dotto/importer.dart';
-import 'package:dotto/components/widgets/progress_indicator.dart';
+import 'package:dotto/widget/progress_indicator.dart';
 import 'package:dotto/feature/my_page/feature/timetable/repository/timetable_repository.dart';
 import 'package:dotto/repository/read_json_file.dart';
